Add CAdjFluidIterion.hpp
