+++ SU2_CFD/include/iteration/CIteration.hpp
@@ -250,57 +250,3 @@ class CIteration {
   virtual void SetRecording(CSolver***** solver, CGeometry**** geometry, CConfig** config, unsigned short val_iZone,
                             unsigned short val_iInst, unsigned short kind_recording) {}
 };
-
-/*!
- * \class CAdjFluidIteration
- * \brief Class for driving an iteration of the adjoint fluid system.
- * \author T. Economon
- */
-class CAdjFluidIteration : public CFluidIteration {
- public:
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CAdjFluidIteration(CConfig* config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CAdjFluidIteration(void) override;
-
-  /*!
-   * \brief Preprocessing to prepare for an iteration of the physics.
-   * \param[in] ??? - Description here.
-   */
-  void Preprocess(COutput* output, CIntegration**** integration, CGeometry**** geometry, CSolver***** solver,
-                  CNumerics****** numerics, CConfig** config, CSurfaceMovement** surface_movement,
-                  CVolumetricMovement*** grid_movement, CFreeFormDefBox*** FFDBox, unsigned short val_iZone,
-                  unsigned short val_iInst) override;
-
-  /*!
-   * \brief Perform a single iteration of the adjoint fluid system.
-   * \param[in] output - Pointer to the COutput class.
-   * \param[in] integration - Container vector with all the integration methods.
-   * \param[in] geometry - Geometrical definition of the problem.
-   * \param[in] solver - Container vector with all the solutions.
-   * \param[in] numerics - Description of the numerical method (the way in which the equations are solved).
-   * \param[in] config - Definition of the particular problem.
-   * \param[in] surface_movement - Surface movement classes of the problem.
-   * \param[in] grid_movement - Volume grid movement classes of the problem.
-   * \param[in] FFDBox - FFD FFDBoxes of the problem.
-   */
-  void Iterate(COutput* output, CIntegration**** integration, CGeometry**** geometry, CSolver***** solver,
-               CNumerics****** numerics, CConfig** config, CSurfaceMovement** surface_movement,
-               CVolumetricMovement*** grid_movement, CFreeFormDefBox*** FFDBox, unsigned short val_iZone,
-               unsigned short val_iInst) override;
-
-  /*!
-   * \brief Updates the containers for the adjoint fluid system.
-   * \param[in] ??? - Description here.
-   */
-  void Update(COutput* output, CIntegration**** integration, CGeometry**** geometry, CSolver***** solver,
-              CNumerics****** numerics, CConfig** config, CSurfaceMovement** surface_movement,
-              CVolumetricMovement*** grid_movement, CFreeFormDefBox*** FFDBox, unsigned short val_iZone,
-              unsigned short val_iInst) override;
-};
